backend/requirements.txt
# Core API
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6
email-validator>=2.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.10.0
psycopg2-binary>=2.9.0

# Configuration & Security
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[argon2]>=1.7.4
slowapi>=0.1.8

# ML & Data
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
shap>=0.41.0

# Utilities
pytz>=2023.3
httpx>=0.24.0
sentry-sdk>=1.20.0

# Development & Testing
pytest
pytest-asyncio
black
isort
flake8
mypy
